requirements.txt
asgi-redis==1.4.2
asgiref==1.1.2
attrs==17.2.0
autobahn==17.6.2
Automat==0.6.0
certifi==2017.4.17
channels==1.1.5
chardet==3.0.4
click==6.7
constantly==15.1.0
coreapi==2.3.1
coreschema==0.0.4
daphne==1.3.0
decorator==4.0.11
Django==1.11.2
django-debug-toolbar==1.8
django-filter==1.0.4
django-smart-selects==1.5.2
djangorestframework==3.6.3
geocoder==1.23.1
gunicorn==19.7.1
hyperlink==17.2.1
idna==2.5
incremental==17.5.0
itypes==1.1.0
Jinja2==2.9.6
MarkupSafe==1.0
msgpack-python==0.4.8
olefile==0.44
Pillow==4.1.1
psycopg2==2.7.1
pytz==2017.2
ratelim==0.1.6
redis==2.10.5
requests==2.18.1
six==1.10.0
sqlparse==0.2.3
Twisted==17.5.0
txaio==2.8.0
uritemplate==3.0.0
urllib3==1.21.1
whitenoise==3.3.0
zope.interface==4.4.2
